add dbUpdateTimeLog in db.models (following last db update)

+++ Api/db/middlewares/db.middlewares.js
@@ -5,20 +5,24 @@ const { compare } = require('../../utils/compare');
 class DbMiddlewares {
   // ********************************** Récupération et copie du fichier Excel
   async storeFile(req, res, next) {
-    const file = req.files.excelFile;
-    const uploadPath = `./uploadedFiles/${file.name}`;
+    if (req.files === null) {
+      res.status(404).send('no file');
+    } else {
+      const file = req.files.excelFile;
+      const uploadPath = `./uploadedFiles/${file.name}`;
 
-    await file.mv(uploadPath, (error) => res.status(500).send(error));
-    // **********************************  Effacement de la copie du fichier Excel après son intégration dans req
-    setTimeout(() => {
-      req.xlsFile = XLSX.readFile('./uploadedFiles/Test.xlsm');
-      const fs = require('fs');
+      await file.mv(uploadPath, (error) => res.status(500).send(error));
+      // **********************************  Effacement de la copie du fichier Excel après son intégration dans req
+      setTimeout(() => {
+        req.xlsFile = XLSX.readFile('./uploadedFiles/Test.xlsm');
+        const fs = require('fs');
 
-      fs.unlink('./uploadedFiles/Test.xlsm', (err) => {
-        if (err) throw err;
-      });
-      next();
-    }, 3000);
+        fs.unlink('./uploadedFiles/Test.xlsm', (err) => {
+          if (err) throw err;
+        });
+        next();
+      }, 3000);
+    }
   }
 
   // ********************************** vérification présence onglet 'Affectation_parc'
